pyproject.toml
[tool.poetry]
name = "Artlyne_gendiff"
version = "0.2.0"
description = "Difference Calculator"
authors = ["Artlyne <Artlynee@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

